Move force kernels back to a common file
