+++ sedov_blast_wave/main.cpp
@@ -1,17 +1,7 @@
 #include "../wash_mockapi.hpp"
-#include "sedov_density_kernel.hpp"
-#include "sedov_eos_kernel.hpp"
-#include "sedov_iad_kernel.hpp"
 #include "sedov_init.hpp"
-#include "sedov_momentum_energy_kernel.hpp"
+#include "sedov_force.hpp"
 #include "sedov_update.hpp"
-
-void force_kernel(wash::Particle& i, const std::vector<wash::Particle>& neighbours) {
-    compute_density(i, neighbours);
-    compute_eos_hydro_std(i);
-    compute_iad(i, neighbours);
-    compute_momentum_energy_std(i, neighbours);
-}
 
 int main(int argc, char **argv) {
     wash::set_precision("double");
