--- sedov_blast_wave/sedov_eos_kernel.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "../wash_mockapi.hpp"
-#include "sedov_consts.hpp"
-
-void compute_eos_hydro_std(wash::Particle& i);
