--- sedov_blast_wave/sedov_momentum_energy_kernel.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "../wash_mockapi.hpp"
-#include "sedov_box.hpp"
-#include "sedov_consts.hpp"
-
-void compute_momentum_energy_std(wash::Particle& i, const std::vector<wash::Particle>& neighbours);
